pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brute-iam"
version = "1.0.0"
description = "Cloud IAM Permission Enumeration Tool for AWS and GCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Brute IAM Contributors" }
]
keywords = [
    "aws",
    "gcp",
    "iam",
    "security",
    "enumeration",
    "permissions",
    "cloud",
    "pentesting",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "click>=8.0.0",
]

[project.optional-dependencies]
aws = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
]
gcp = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
]
all = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
iamx = "iamx.cli:main"

[project.urls]
Homepage = "https://github.com/brute-iam/brute-iam"
Documentation = "https://github.com/brute-iam/brute-iam#readme"
Repository = "https://github.com/brute-iam/brute-iam"
Issues = "https://github.com/brute-iam/brute-iam/issues"

[tool.hatch.build.targets.wheel]
packages = ["iamx"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hatch
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["iamx"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
